deduplicate auth e2e assertions
