+++ e2e/auth-integration.spec.ts
@@ -1,5 +1,6 @@
 import { test, expect } from "@playwright/test";
 import { existsSync } from "node:fs";
+import { exerciseConfiguredModeSwitching } from "./auth-surface";
 
 const hasAuth = existsSync("playwright/.auth/user.json");
 
@@ -30,23 +31,6 @@ test.describe("authenticated auth integration", () => {
     test.skip(!hasAuth, "No auth state found. Run `pnpm auth:local` first.");
 
     await page.goto("/login");
-    const authPage = page.getByRole("main");
-
-    await expect(authPage.getByRole("heading", { name: "Sign in" })).toBeVisible();
-    await expect(authPage.getByRole("button", { name: "Continue with Google" })).toBeVisible();
-    await expect(authPage.getByLabel("Email address")).toBeVisible();
-    await expect(authPage.getByLabel("Password")).toBeVisible();
-
-    await authPage.getByRole("button", { name: "Create account" }).click();
-    await expect(authPage.getByRole("heading", { name: "Create account" })).toBeVisible();
-    await expect(authPage.getByLabel("Confirm password")).toBeVisible();
-
-    await authPage.getByRole("button", { name: /already have an account\? sign in/i }).click();
-    await expect(authPage.getByRole("heading", { name: "Sign in" })).toBeVisible();
-
-    await authPage.getByRole("button", { name: "Forgot password?" }).click();
-    await expect(authPage.getByRole("heading", { name: "Reset password" })).toBeVisible();
-    await expect(authPage.getByRole("button", { name: "Send reset link" })).toBeVisible();
-    await expect(authPage.getByLabel("Password")).toHaveCount(0);
+    await exerciseConfiguredModeSwitching(page);
   });
 });
